Provide option to sort the arrow tables
Previously, the arrow endpoints are not guaranteed to present their data
in a sorted order, this PR lets the user set that for single stream
calls, and sorts by time for the streamset operations by default.

Streamset transformers like streamset.to_dataframe in the old version of
the api used a PointBuffer that would sort the values by time before
returning to the user.

The single stream arrow methods now have a boolean argument `sorted` to
let the user specify if they want to sort the returned table on the
'time' column or not.
This is False by default.

The streamset methods though, will be sorted by default and the user
wont be able to switch that. We can change that later if needed.