Remove 4MB limit for gRPC message payloads (#49)
Currently our logic chunks data to send over gRPC according to numbers
of rows of data, this led to large streamsets in the multivalue api
erroring out while trying to recv this data.

This PR sets the limit for the client to receive from the server to be
unlimited for the time being. This will allow arbitrarly-sized
streamsets to have successful multivalue queries.

In the future we should update our logic to better handle these size
limits when sending from the server, but this is a patch fix for now.